requirements.txt
# Core FastAPI and Server Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Audio Processing & Speech Recognition
openai-whisper==20231117
ffmpeg-python==0.2.0
pydub==0.25.1
soundfile==0.12.1
scipy==1.11.4

# Text-to-Speech (TTS)
TTS==0.22.0

# LLM & Embeddings
langchain==0.1.0
langchain-community==0.0.12
langchain-ollama==0.0.2
langchain-text-splitters==0.0.1
groq==0.4.1
ollama==0.1.7

# Vector Store & Retrieval
chromadb==0.4.22
sentence-transformers==2.2.2
rank-bm25==0.2.2

# Machine Learning
torch==2.1.2
transformers==4.36.2
numpy==1.24.4

# Document Processing
pypdf==3.17.4
PyMuPDF==1.23.8

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
PyYAML==6.0.1

# Web Interface
streamlit==1.29.0

# Utilities
tqdm==4.66.1

# Development Dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0
mypy==1.8.0

# Additional dependencies for production
gunicorn==21.2.0
python-json-logger==2.0.7
prometheus-client==0.19.0
